==> 1.get_filers_from_year_city.ipynb <==
# Get a subset of a state (by filtering for a city)

## Download the 2017 metadata and show where we put it

Use irsx_index at the command line to retrieve the 2017 listing of all xml 990 efilings.

__Note that these are filings received *during* 2017, so check the tax_period__

Irsx_index is a helper command that is included with irsx, so you need to have installed it first. Try `pip install irsx` or see more [here](https://github.com/jsfenfen/990-xml-reader/#installation).

We use the `--verbose` flag so can watch it's progress

    $ irsx_index --verbose --year=2017
    Getting index file for year: 2017 
    remote=https://s3.amazonaws.com/irs-form-990/index_2017.csv 
    local=/Users/jfenton/github-whitelabel/envs/irs-cookbook/lib/python3.6/site-packages/irsx/CSV/index_2017.csv
    Beginning streaming download of https://s3.amazonaws.com/irs-form-990/index_2017.csv
    Total file size: 59.45 MB

## You don't need to run the command below if you've run irs_index at the command line
## To actually do this from within the notebook environment uncomment the below
## Note that we're using the %sx 'magic command' which captures the output as an array 
## Your mileage may vary depending on how jupyter plays with your operating system

# %sx irsx_index --verbose --year=2017

# Get all nonprofit organizations in your state

We grabbed a file for just the state of Oregon as eo_or.csv from here: 
https://www.irs.gov/charities-non-profits/exempt-organizations-business-master-file-extract-eo-bmf

Note that this method isn't great for historic organizations; the IRS purges organizations after they've become inactive for a period of time. Historic EO BMF files are available here: http://nccs-data.urban.org/data.php?ds=bmf 

# importing libraries we'll use.
import csv
import os
import pandas as pd

# This tells us where the csv files are located in the system
from irsx.settings import INDEX_DIRECTORY

oregon_np = pd.read_csv("eo_or.csv")

# look at the first few lines
#oregon_np.head()
## print the headers as an array
list(oregon_np)

# Ignore some columns for now
or_np_simplified = oregon_np.filter(items=['EIN', 'NAME', 'ICO', 'STREET', 'CITY', 'STATE', 'ZIP', 'INCOME_AMT', 'ASSET_AMT', 'TAX_PERIOD'])
print("total oregon orgs: %s" % len(or_np_simplified))

# This is a toy filter for a demo -- you'd want something more robust than a perfect text match
pdx_orgs = or_np_simplified.query('CITY == "PORTLAND"')
print("total Portland, OR orgs: %s" % len(pdx_orgs))

# Show the top values by income 
pdx_orgs.sort_values(by=['INCOME_AMT'], ascending=[0]).head()



# this is from the index file we dowloaded at the start
INDEX_2017= os.path.join(INDEX_DIRECTORY, 'index_2017.csv')
np_2017 = pd.read_csv(INDEX_2017)


## Now save the list of possible filers who actually filed out to .csv

# Now find orgs that are in portland that filed in 2017. 
# This join requires that both fields be named EIN and be formatted the same

pdx_2017_efilers = pd.merge(np_2017,
                 pdx_orgs,
                 on='EIN')
print("Found a total of %s portland 2017 efilers" % len(pdx_2017_efilers))
pdx_2017_efilers.head()

# sort by income amt, asset amt
pdx_2017_efilers = pdx_2017_efilers.sort_values(by=['INCOME_AMT', 'ASSET_AMT'], ascending=[0,0])
# Lets write them back out to a file for reference.
pdx_2017_efilers.to_csv('pdxefilers.csv')

# These are the top few for reference
pdx_2017_efilers.head()
